requirements.txt
Django
djangorestframework
django-cors-headers
dj-database-url
python-dotenv
psycopg[binary]
rav
python-decouple
gunicorn

PyPDF2==3.0.1
python-docx==0.8.11
odfpy==1.4.1
nltk==3.8.1
requests==2.31.0



# requirements.txt

openai
pandas
psycopg[binary]
pgvector
python-dotenv
python-docx
odfpy   

pymupdf


python-magic-bin
unstructured[local-inference]
unstructured-inference
lxml



# PDF Processing
PyMuPDF==1.23.8
pdf2image==1.16.3
pytesseract==0.3.10
PyPDF2==3.0.1
 pdfminer.six

# Document Processing
python-docx==0.8.11
python-pptx==0.6.21
unstructured==0.12.0
pandas==2.1.4
numpy==1.26.3

# OCR and Image Processing
Pillow==10.2.0
pytesseract==0.3.10
